test(spring-boot-starter-azure-grpc-spring): add context loading and logging tests
